+++ dataline-scheduler/src/main/java/io/dataline/scheduler/WorkerRunner.java
@@ -30,6 +30,7 @@ import io.dataline.config.JobSyncConfig;
 import io.dataline.config.StandardCheckConnectionInput;
 import io.dataline.config.StandardDiscoverSchemaInput;
 import io.dataline.config.StandardSyncInput;
+import io.dataline.scheduler.persistence.SchedulerPersistence;
 import io.dataline.workers.DefaultSyncWorker;
 import io.dataline.workers.process.ProcessBuilderFactory;
 import io.dataline.workers.singer.SingerCheckConnectionWorker;
