+++ dataline-scheduler/src/test/java/io/dataline/scheduler/JobSchedulerTest.java
@@ -46,11 +46,12 @@ import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigPersistence;
 import io.dataline.config.persistence.JsonValidationException;
 import io.dataline.integrations.Integrations;
-import io.dataline.scheduler.job_creation.SyncJobFactory;
+import io.dataline.scheduler.job_factory.SyncJobFactory;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.UUID;
+import io.dataline.scheduler.persistence.SchedulerPersistence;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
